fixed FreeImage_Rotate() to work with images over 5000 mgpx; i changed unsigned to int64 types

+++ Source/FreeImageToolkit/ClassicRotate.cpp
@@ -53,10 +53,10 @@ Parameter T can be BYTE, WORD of float.
 */
 template <class T> void 
 HorizontalSkewT(FIBITMAP *src, FIBITMAP *dst, int row, int iOffset, double weight, const void *bkcolor = NULL) {
-	int iXPos;
+	INT64 iXPos;
 
-	const unsigned src_width  = FreeImage_GetWidth(src);
-	const unsigned dst_width  = FreeImage_GetWidth(dst);
+	const INT64 src_width  = FreeImage_GetWidth(src);
+	const INT64 dst_width  = FreeImage_GetWidth(dst);
 
 	T pxlSrc[4], pxlLeft[4], pxlOldLeft[4];	// 4 = 4*sizeof(T) max
 	
@@ -69,9 +69,9 @@ HorizontalSkewT(FIBITMAP *src, FIBITMAP *dst, int row, int iOffset, double weigh
 	}
 
 	// calculate the number of bytes per pixel
-	const unsigned bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
+	const INT64 bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
 	// calculate the number of samples per pixel
-	const unsigned samples = bytespp / sizeof(T);
+	const INT64 samples = bytespp / sizeof(T);
 
 	BYTE *src_bits = FreeImage_GetScanLine(src, row);
 	BYTE *dst_bits = FreeImage_GetScanLine(dst, row);
@@ -89,18 +89,18 @@ HorizontalSkewT(FIBITMAP *src, FIBITMAP *dst, int row, int iOffset, double weigh
 		memset(&pxlOldLeft[0], 0, bytespp);
 	}
 
-	for(unsigned i = 0; i < src_width; i++) {
+	for(INT64 i = 0; i < src_width; i++) {
 		// loop through row pixels
 		AssignPixel((BYTE*)&pxlSrc[0], (BYTE*)src_bits, bytespp);
 		// calculate weights
-		for(unsigned j = 0; j < samples; j++) {
+		for(INT64 j = 0; j < samples; j++) {
 			pxlLeft[j] = static_cast<T>(pxlBkg[j] + (pxlSrc[j] - pxlBkg[j]) * weight + 0.5);
 		}
 		// check boundaries 
 		iXPos = i + iOffset;
 		if((iXPos >= 0) && (iXPos < (int)dst_width)) {
 			// update left over on source
-			for(unsigned j = 0; j < samples; j++) {
+			for(INT64 j = 0; j < samples; j++) {
 				pxlSrc[j] = pxlSrc[j] - (pxlLeft[j] - pxlOldLeft[j]);
 			}
 			AssignPixel((BYTE*)&dst_bits[iXPos*bytespp], (BYTE*)&pxlSrc[0], bytespp);
@@ -124,7 +124,7 @@ HorizontalSkewT(FIBITMAP *src, FIBITMAP *dst, int row, int iOffset, double weigh
 		// clear to the right of the skewed line with background
 		dst_bits += bytespp;
 		if(bkcolor) {
-			for(unsigned i = 0; i < dst_width - iXPos - 1; i++) {
+			for(INT64 i = 0; i < dst_width - iXPos - 1; i++) {
 				memcpy(&dst_bits[i * bytespp], bkcolor, bytespp);
 			}
 		} else {
@@ -184,10 +184,10 @@ Parameter T can be BYTE, WORD of float.
 */
 template <class T> void 
 VerticalSkewT(FIBITMAP *src, FIBITMAP *dst, int col, int iOffset, double weight, const void *bkcolor = NULL) {
-	int iYPos;
+	INT64 iYPos;
 
-	unsigned src_height = FreeImage_GetHeight(src);
-	unsigned dst_height = FreeImage_GetHeight(dst);
+	INT64 src_height = FreeImage_GetHeight(src);
+	INT64 dst_height = FreeImage_GetHeight(dst);
 
 	T pxlSrc[4], pxlLeft[4], pxlOldLeft[4];	// 4 = 4*sizeof(T) max
 
@@ -200,44 +200,44 @@ VerticalSkewT(FIBITMAP *src, FIBITMAP *dst, int col, int iOffset, double weight,
 	}
 
 	// calculate the number of bytes per pixel
-	const unsigned bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
+	const INT64 bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
 	// calculate the number of samples per pixel
-	const unsigned samples = bytespp / sizeof(T);
+	const INT64 samples = bytespp / sizeof(T);
 
-	const unsigned src_pitch = FreeImage_GetPitch(src);
-	const unsigned dst_pitch = FreeImage_GetPitch(dst);
-	const unsigned index = col * bytespp;
+	const INT64 src_pitch = FreeImage_GetPitch(src);
+	const INT64 dst_pitch = FreeImage_GetPitch(dst);
+	const INT64 index = col * bytespp;
 
 	BYTE *src_bits = FreeImage_GetBits(src) + index;
 	BYTE *dst_bits = FreeImage_GetBits(dst) + index;
 
 	// fill gap above skew with background
 	if(bkcolor) {
-		for(int k = 0; k < iOffset; k++) {
+		for(INT64 k = 0; k < iOffset; k++) {
 			memcpy(dst_bits, bkcolor, bytespp);
 			dst_bits += dst_pitch;
 		}
 		memcpy(&pxlOldLeft[0], bkcolor, bytespp);
 	} else {
-		for(int k = 0; k < iOffset; k++) {
+		for(INT64 k = 0; k < iOffset; k++) {
 			memset(dst_bits, 0, bytespp);
 			dst_bits += dst_pitch;
 		}
 		memset(&pxlOldLeft[0], 0, bytespp);
 	}
 
-	for(unsigned i = 0; i < src_height; i++) {
+	for(INT64 i = 0; i < src_height; i++) {
 		// loop through column pixels
 		AssignPixel((BYTE*)(&pxlSrc[0]), src_bits, bytespp);
 		// calculate weights
-		for(unsigned j = 0; j < samples; j++) {
+		for(INT64 j = 0; j < samples; j++) {
 			pxlLeft[j] = static_cast<T>(pxlBkg[j] + (pxlSrc[j] - pxlBkg[j]) * weight + 0.5);
 		}
 		// check boundaries
 		iYPos = i + iOffset;
 		if((iYPos >= 0) && (iYPos < (int)dst_height)) {
 			// update left over on source
-			for(unsigned j = 0; j < samples; j++) {
+			for(INT64 j = 0; j < samples; j++) {
 				pxlSrc[j] = pxlSrc[j] - (pxlLeft[j] - pxlOldLeft[j]);
 			}
 			dst_bits = FreeImage_GetScanLine(dst, iYPos) + index;
@@ -322,10 +322,10 @@ Rotate90(FIBITMAP *src) {
 
 	const unsigned bpp = FreeImage_GetBPP(src);
 
-	const unsigned src_width  = FreeImage_GetWidth(src);
-	const unsigned src_height = FreeImage_GetHeight(src);	
-	const unsigned dst_width  = src_height;
-	const unsigned dst_height = src_width;
+	const INT64 src_width  = FreeImage_GetWidth(src);
+	const INT64 src_height = FreeImage_GetHeight(src);	
+	const INT64 dst_width  = src_height;
+	const INT64 dst_height = src_width;
 
 	FREE_IMAGE_TYPE image_type = FreeImage_GetImageType(src);
 
@@ -334,8 +334,8 @@ Rotate90(FIBITMAP *src) {
 	if(NULL == dst) return NULL;
 
 	// get src and dst scan width
-	const unsigned src_pitch  = FreeImage_GetPitch(src);
-	const unsigned dst_pitch  = FreeImage_GetPitch(dst);
+	const INT64 src_pitch  = FreeImage_GetPitch(src);
+	const INT64 dst_pitch  = FreeImage_GetPitch(dst);
 
 	switch(image_type) {
 		case FIT_BITMAP:
@@ -344,16 +344,15 @@ Rotate90(FIBITMAP *src) {
 
 				BYTE *bsrc  = FreeImage_GetBits(src); 
 				BYTE *bdest = FreeImage_GetBits(dst);
-
 				BYTE *dbitsmax = bdest + dst_height * dst_pitch - 1;
-
-				for(unsigned y = 0; y < src_height; y++) {
+				
+				for(INT64 y = 0; y < src_height; y++) {
 					// figure out the column we are going to be copying to
-					const div_t div_r = div(y, 8);
+					const div_t div_r = div((int)y, 8);
 					// set bit pos of src column byte
 					const BYTE bitpos = (BYTE)(128 >> div_r.rem);
 					BYTE *srcdisp = bsrc + y * src_pitch;
-					for(unsigned x = 0; x < src_pitch; x++) {
+					for(INT64 x = 0; x < src_pitch; x++) {
 						// get source bits
 						BYTE *sbits = srcdisp + x;
 						// get destination column
@@ -380,21 +379,22 @@ Rotate90(FIBITMAP *src) {
 				BYTE *bdest = FreeImage_GetBits(dst);  // destination pixels
 
 				// calculate the number of bytes per pixel (1 for 8-bit, 3 for 24-bit or 4 for 32-bit)
-				const unsigned bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
+				const INT64 bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
 				
 				// for all image blocks of RBLOCK*RBLOCK pixels
 				
 				// x-segment
-				for(unsigned xs = 0; xs < dst_width; xs += RBLOCK) {
+				
+				for(INT64 xs = 0; xs < dst_width; xs += RBLOCK) {
 					// y-segment
-					for(unsigned ys = 0; ys < dst_height; ys += RBLOCK) {
-						for(unsigned y = ys; y < MIN(dst_height, ys + RBLOCK); y++) {    // do rotation
-							const unsigned y2 = dst_height - y - 1;
+					for(INT64 ys = 0; ys < dst_height; ys += RBLOCK) {
+						for(INT64 y = ys; y < MIN(dst_height, ys + RBLOCK); y++) {    // do rotation
+							const INT64 y2 = dst_height - y - 1;
 							// point to src pixel at (y2, xs)
 							BYTE *src_bits = bsrc + (xs * src_pitch) + (y2 * bytespp);
 							// point to dst pixel at (xs, y)
 							BYTE *dst_bits = bdest + (y * dst_pitch) + (xs * bytespp);
-							for(unsigned x = xs; x < MIN(dst_width, xs + RBLOCK); x++) {
+							for(INT64 x = xs; x < MIN(dst_width, xs + RBLOCK); x++) {
 								// dst.SetPixel(x, y, src.GetPixel(y2, x));
 								AssignPixel(dst_bits, src_bits, bytespp);
 								dst_bits += bytespp;
@@ -416,12 +416,12 @@ Rotate90(FIBITMAP *src) {
 			BYTE *bdest = FreeImage_GetBits(dst);  // destination pixels
 
 			// calculate the number of bytes per pixel
-			const unsigned bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
-
-			for(unsigned y = 0; y < dst_height; y++) {
+			const INT64 bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
+			
+			for(INT64 y = 0; y < dst_height; y++) {
 				BYTE *src_bits = bsrc + (src_width - 1 - y) * bytespp;
 				BYTE *dst_bits = bdest + (y * dst_pitch);
-				for(unsigned x = 0; x < dst_width; x++) {
+				for(INT64 x = 0; x < dst_width; x++) {
 					AssignPixel(dst_bits, src_bits, bytespp);
 					src_bits += src_pitch;
 					dst_bits += bytespp;
@@ -442,14 +442,13 @@ Precise rotation, no filters required.
 */
 static FIBITMAP* 
 Rotate180(FIBITMAP *src) {
-	int x, y, k, pos;
+	INT64 x, y, k, pos;
 
 	const int bpp = FreeImage_GetBPP(src);
-
-	const int src_width  = FreeImage_GetWidth(src);
-	const int src_height = FreeImage_GetHeight(src);
-	const int dst_width  = src_width;
-	const int dst_height = src_height;
+	const INT64 src_width  = FreeImage_GetWidth(src);
+	const INT64 src_height = FreeImage_GetHeight(src);
+	const INT64 dst_width  = src_width;
+	const INT64 dst_height = src_height;
 
 	FREE_IMAGE_TYPE image_type = FreeImage_GetImageType(src);
 
@@ -459,10 +458,11 @@ Rotate180(FIBITMAP *src) {
 	switch(image_type) {
 		case FIT_BITMAP:
 			if(bpp == 1) {
-				for(int y = 0; y < src_height; y++) {
+				
+				for(INT64 y = 0; y < src_height; y++) {
 					BYTE *src_bits = FreeImage_GetScanLine(src, y);
 					BYTE *dst_bits = FreeImage_GetScanLine(dst, dst_height - y - 1);
-					for(int x = 0; x < src_width; x++) {
+					for(INT64 x = 0; x < src_width; x++) {
 						// get bit at (x, y)
 						k = (src_bits[x >> 3] & (0x80 >> (x & 0x07))) != 0;
 						// set bit at (dst_width - x - 1, dst_height - y - 1)
@@ -481,8 +481,8 @@ Rotate180(FIBITMAP *src) {
 		case FIT_RGBAF:
 		{
 			 // Calculate the number of bytes per pixel
-			const int bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
-
+			const INT64 bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
+			
 			for(y = 0; y < src_height; y++) {
 				BYTE *src_bits = FreeImage_GetScanLine(src, y);
 				BYTE *dst_bits = FreeImage_GetScanLine(dst, dst_height - y - 1) + (dst_width - 1) * bytespp;
@@ -510,14 +510,14 @@ Code adapted from CxImage (http://www.xdp.it/cximage.htm)
 */
 static FIBITMAP* 
 Rotate270(FIBITMAP *src) {
-	int x2, dlineup;
+	INT64 x2, dlineup;
 
 	const unsigned bpp = FreeImage_GetBPP(src);
 
-	const unsigned src_width  = FreeImage_GetWidth(src);
-	const unsigned src_height = FreeImage_GetHeight(src);	
-	const unsigned dst_width  = src_height;
-	const unsigned dst_height = src_width;
+	const INT64 src_width  = FreeImage_GetWidth(src);
+	const INT64 src_height = FreeImage_GetHeight(src);	
+	const INT64 dst_width  = src_height;
+	const INT64 dst_height = src_width;
 
 	FREE_IMAGE_TYPE image_type = FreeImage_GetImageType(src);
 
@@ -526,8 +526,8 @@ Rotate270(FIBITMAP *src) {
 	if(NULL == dst) return NULL;
 
 	// get src and dst scan width
-	const unsigned src_pitch  = FreeImage_GetPitch(src);
-	const unsigned dst_pitch  = FreeImage_GetPitch(dst);
+	const INT64 src_pitch  = FreeImage_GetPitch(src);
+	const INT64 dst_pitch  = FreeImage_GetPitch(dst);
 	
 	switch(image_type) {
 		case FIT_BITMAP:
@@ -538,19 +538,19 @@ Rotate270(FIBITMAP *src) {
 				BYTE *bdest = FreeImage_GetBits(dst);
 				BYTE *dbitsmax = bdest + dst_height * dst_pitch - 1;
 				dlineup = 8 * dst_pitch - dst_width;
-
-				for(unsigned y = 0; y < src_height; y++) {
+				
+				for(INT64 y = 0; y < src_height; y++) {
 					// figure out the column we are going to be copying to
-					const div_t div_r = div(y + dlineup, 8);
+					const div_t div_r = div((int)(y + dlineup), 8);
 					// set bit pos of src column byte
 					const BYTE bitpos = (BYTE)(1 << div_r.rem);
 					const BYTE *srcdisp = bsrc + y * src_pitch;
-					for(unsigned x = 0; x < src_pitch; x++) {
+					for(INT64 x = 0; x < src_pitch; x++) {
 						// get source bits
 						const BYTE *sbits = srcdisp + x;
 						// get destination column
 						BYTE *nrow = bdest + (x * 8) * dst_pitch + dst_pitch - 1 - div_r.quot;
-						for(unsigned z = 0; z < 8; z++) {
+						for(INT64 z = 0; z < 8; z++) {
 						   // get destination byte
 							BYTE *dbits = nrow + z * dst_pitch;
 							if ((dbits < bdest) || (dbits > dbitsmax)) break;
@@ -572,21 +572,22 @@ Rotate270(FIBITMAP *src) {
 				BYTE *bdest = FreeImage_GetBits(dst);  // destination pixels
 
 				// Calculate the number of bytes per pixel (1 for 8-bit, 3 for 24-bit or 4 for 32-bit)
-				const unsigned bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
+				const INT64 bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
 
 				// for all image blocks of RBLOCK*RBLOCK pixels
 
 				// x-segment
-				for(unsigned xs = 0; xs < dst_width; xs += RBLOCK) {
+				
+				for(INT64 xs = 0; xs < dst_width; xs += RBLOCK) {
 					// y-segment
-					for(unsigned ys = 0; ys < dst_height; ys += RBLOCK) {
-						for(unsigned x = xs; x < MIN(dst_width, xs + RBLOCK); x++) {    // do rotation
+					for(INT64 ys = 0; ys < dst_height; ys += RBLOCK) {
+						for(INT64 x = xs; x < MIN(dst_width, xs + RBLOCK); x++) {    // do rotation
 							x2 = dst_width - x - 1;
 							// point to src pixel at (ys, x2)
 							BYTE *src_bits = bsrc + (x2 * src_pitch) + (ys * bytespp);
 							// point to dst pixel at (x, ys)
 							BYTE *dst_bits = bdest + (ys * dst_pitch) + (x * bytespp);
-							for(unsigned y = ys; y < MIN(dst_height, ys + RBLOCK); y++) {
+							for(INT64 y = ys; y < MIN(dst_height, ys + RBLOCK); y++) {
 								// dst.SetPixel(x, y, src.GetPixel(y, x2));
 								AssignPixel(dst_bits, src_bits, bytespp);
 								src_bits += bytespp;
@@ -608,12 +609,12 @@ Rotate270(FIBITMAP *src) {
 			BYTE *bdest = FreeImage_GetBits(dst);  // destination pixels
 
 			// calculate the number of bytes per pixel
-			const unsigned bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
-
-			for(unsigned y = 0; y < dst_height; y++) {
+			const INT64 bytespp = FreeImage_GetLine(src) / FreeImage_GetWidth(src);
+			
+			for(INT64 y = 0; y < dst_height; y++) {
 				BYTE *src_bits = bsrc + (src_height - 1) * src_pitch + y * bytespp;
 				BYTE *dst_bits = bdest + (y * dst_pitch);
-				for(unsigned x = 0; x < dst_width; x++) {
+				for(INT64 x = 0; x < dst_width; x++) {
 					AssignPixel(dst_bits, src_bits, bytespp);
 					src_bits -= src_pitch;
 					dst_bits += bytespp;
@@ -637,7 +638,7 @@ static FIBITMAP*
 Rotate45(FIBITMAP *src, double dAngle, const void *bkcolor) {
 	const double ROTATE_PI = double(3.1415926535897932384626433832795);
 
-	unsigned u;
+	INT64 u;
 
 	const unsigned bpp = FreeImage_GetBPP(src);
 
@@ -645,14 +646,14 @@ Rotate45(FIBITMAP *src, double dAngle, const void *bkcolor) {
 	const double dSinE = sin(dRadAngle);
 	const double dTan = tan(dRadAngle / 2);
 
-	const unsigned src_width  = FreeImage_GetWidth(src);
-	const unsigned src_height = FreeImage_GetHeight(src);
+	const INT64 src_width  = FreeImage_GetWidth(src);
+	const INT64 src_height = FreeImage_GetHeight(src);
 
 	FREE_IMAGE_TYPE image_type = FreeImage_GetImageType(src);
 
 	// Calc first shear (horizontal) destination image dimensions 
-	const unsigned width_1  = src_width + unsigned((double)src_height * fabs(dTan) + 0.5);
-	const unsigned height_1 = src_height; 
+	const INT64 width_1  = src_width + unsigned((double)src_height * fabs(dTan) + 0.5);
+	const INT64 height_1 = src_height; 
 
 	// Perform 1st shear (horizontal)
 	// ----------------------------------------------------------------------
@@ -682,8 +683,8 @@ Rotate45(FIBITMAP *src, double dAngle, const void *bkcolor) {
 	// ----------------------------------------------------------------------
 
 	// Calc 2nd shear (vertical) destination image dimensions
-	const unsigned width_2  = width_1;
-	unsigned height_2 = unsigned((double)src_width * fabs(dSinE) + (double)src_height * cos(dRadAngle) + 0.5) + 1;
+	const INT64 width_2  = width_1;
+	INT64 height_2 = unsigned((double)src_width * fabs(dSinE) + (double)src_height * cos(dRadAngle) + 0.5) + 1;
 
 	// Allocate image for 2nd shear
 	FIBITMAP *dst2 = FreeImage_AllocateT(image_type, width_2, height_2, bpp);
@@ -714,8 +715,8 @@ Rotate45(FIBITMAP *src, double dAngle, const void *bkcolor) {
 	FreeImage_Unload(dst1);
 
 	// Calc 3rd shear (horizontal) destination image dimensions
-	const unsigned width_3  = unsigned(double(src_height) * fabs(dSinE) + double(src_width) * cos(dRadAngle) + 0.5) + 1;
-	const unsigned height_3 = height_2;
+	const INT64 width_3  = unsigned(double(src_height) * fabs(dSinE) + double(src_width) * cos(dRadAngle) + 0.5) + 1;
+	const INT64 height_3 = height_2;
 
 	// Allocate image for 3rd shear
 	FIBITMAP *dst3 = FreeImage_AllocateT(image_type, width_3, height_3, bpp);
